Add option to jump if already open and/or open in new tab
This commit adds two settings:

`g:fuzzy_tabopen` opens the file in a new tab instead of the current
one, it otherwise behaves the same as before.

`g:fuzzy_jump_if_open` jumps to the file if it is already open in
another tab/buffer. It uses the `:drop` command to accomplish this.

If you combine both of these options you have the following situation:

- if the file is not yet opened in vim: opens file in new tab
- if the file is already opened in vim: jump to the file

If only `g:fuzzy_tabopen` is set:

- if the file is not yet opened in vim: opens file in new tab
- if the file is already opened in vim: open the file in a new tab

If only `g:fuzzy_jump_if_open` is set:

- if the file is not yet opened in vim: opens file in current window
- if the file is already opened in vim: jump to the file

If neither is set, the old behaviour is used:

- if the file is not yet opened in vim: opens file in current window
- if the file is already opened in vim: opens file in current window