Pipe lines directly to `fzy`
Old behavior was loading all the commands into a Vim list, write it into
a file, and pipe that file into `fzy`.

This is a inefficient, because both `ag`/`rg` and `fzy` are blazing fast
and can handle large quantities of data, but Vimscript, with it's
dynamic typing and garbage-collected memory - which aren't even
state-of-the-art! - is not only slowing the entire process down but also
breaks it when doing a big `:FuzzyGrep` in a large codebases.

This PR aims to fix it:

* With `:FuzzyGrep` it pipes the `ag`/`rg` command directly to `fzy`.
* With `:FuzzyOpen` it creates a temporary file, writes the list of
  buffers to it with `writefile()` and then uses `ag`/`rg` to pipe the
  list of files into that temporary file without passing them through
  Vim.

This ensures that aside from the list of buffers(that for obvious
reasons must come from Vim) all the match choices are piped by the
operation system and Vim does not need to load them into it's memory.